More setMaxRows, promote ids to long, more PreparedStatement, others
The project_id is a long in the db, promote it in the code.
The timeseries_id is still int in the db, but eventually should be
long, prepare the code for that.

Use static Set instances when looking up rdbms capability support.

Fixes #94458 (PreparedStatement use, setMaxRows)
Fixes #93877 (setMaxRows)
Refs #60801 (Promotion of some ids to long)

+++ wres-io/src/wres/io/data/caching/Ensembles.java
@@ -60,7 +60,7 @@ public class Ensembles extends Cache<EnsembleDetails, String> {
         while (data.next()) {
             detail = new EnsembleDetails();
             detail.setEnsembleName(data.getString("ensemble_name"));
-            detail.setID(data.getInt("ensemble_id"));
+            detail.setID( data.getLong( "ensemble_id" ) );
             this.add( detail );
         }
     }
