+++ wres-io/test/wres/io/retrieval/RetrieverTestConstants.java
@@ -3,7 +3,7 @@ package wres.io.retrieval;
 class RetrieverTestConstants
 {
     /** A project surrogate id for testing */
-    static final Integer PROJECT_ID = 1;
+    static final long PROJECT_ID = 1;
 
     /** A project natural id (hash) for testing */
     static final String PROJECT_HASH = "123deadbeef456";
